+++ delicatessen/main.py
@@ -1,3 +1,6 @@
+# delicatessen
+from . import modules
+
 # Standard library
 import pathlib
 import sys
@@ -85,14 +88,14 @@ class Selector:
                 self.widget,
                 *additional_widgets,
                 width=width,
-                css_classes=["controls"]
+                css_classes=["controls"],
             ),
             footer,
             css_classes=self.css_classes,
         )
 
 
-class PrimaryPlot:
+class Plot:
     def __init__(self, dataset, parameters):
 
         self.dataset = dataset
@@ -196,6 +199,9 @@ class PrimaryPlot:
         ]:
             control.widget.on_change("value", self.callback)
 
+        # Load and display the data
+        self.callback(None, None, None)
+
     def callback(self, attr, old, new):
         """
         Triggered when the user changes what we're plotting on the main plot.
@@ -230,61 +236,38 @@ class PrimaryPlot:
             color=color,
         )
 
-
-class SecondaryPlot:
-    def __init__(self, primary_plot):
-        self.primary_plot = primary_plot
-        self.source = ColumnDataSource(data=dict(x=[], y=[]))
-        self.plot = figure(
-            plot_height=300, plot_width=700, title="", sizing_mode="scale_both"
+    def layout(self):
+        inputs_left = column(
+            self.data.layout(),
+            Spacer(height=10),
+            self.specials.layout(),
+            width=160,
         )
-        self.plot.circle(
-            x="x",
-            y="y",
-            source=self.source,
-            line_color=None,
-            color="black",
-            alpha=0.1,
+        inputs_right = column(
+            self.xaxis.layout([self.yaxis.widget]),
+            Spacer(height=10),
+            self.size.layout([self.color.widget]),
         )
-
-        # Register the callback
-        self.primary_plot.source.selected.on_change("indices", self.callback)
-
-    def callback(self, attr, old, new):
-        """
-        Triggered when the user selects a point on the main plot.
-
-        """
-        # If a point is selected...
-        if len(self.primary_plot.source.selected.indices):
-
-            # Get the TIC ID
-            ticid = self.primary_plot.source.data["ticid"][
-                self.primary_plot.source.selected.indices[0]
-            ]
-            print("Fetching data for TIC ID {0}".format(ticid))
-
-            # TODO: Actually fetch the data from MAST.
-            # For now just populate with random numbers
-            self.source.data = dict(
-                x=np.linspace(0, 1, 10000), y=np.random.randn(10000)
-            )
-
-        else:
-
-            # Clear the plot
-            self.source.data = dict(x=[], y=[])
+        header = Div(
+            text=f"""<img src="{LOGO_URL}"></img>""",
+            css_classes=["header-image"],
+            width=320,
+            height=100,
+        )
+        inputs = column(
+            header, row(inputs_left, Spacer(width=10), inputs_right)
+        )
+        return row(inputs, Spacer(width=10), self.plot)
 
 
 class Delicatessen:
-
     def __init__(self, data_file=None):
 
         # This is to have a default / test data file to show. But we probably
         # want to change this, or remove the default when we "release"!
         if data_file is None:
             deli_path = pathlib.Path(__file__).parent.absolute()
-            data_file = deli_path / 'data' / 'TESS-Gaia-mini.csv'
+            data_file = deli_path / "data" / "TESS-Gaia-mini.csv"
 
         # The data file can be any file format that astropy.table can read:
         data = at.Table.read(data_file)
@@ -297,34 +280,11 @@ class Delicatessen:
         self.dataset = dataset
 
         # Instantiate the plots
-        self.primary = PrimaryPlot(dataset, parameters)
-        self.secondary = SecondaryPlot(self.primary)
+        self.primary = Plot(dataset, parameters)
+        self.secondary = modules.TESSGaiaMini.Plot(self.primary)
 
         # Display things on the page
-        inputs_left = column(
-            self.primary.data.layout(),
-            Spacer(height=10),
-            self.primary.specials.layout(),
-            width=160,
-        )
-        inputs_right = column(
-            self.primary.xaxis.layout([self.primary.yaxis.widget]),
-            Spacer(height=10),
-            self.primary.size.layout([self.primary.color.widget]),
-        )
-        header = Div(
-            text=f"""<img src="{LOGO_URL}"></img>""",
-            css_classes=["header-image"],
-            width=320,
-            height=100,
-        )
-        inputs = column(header,
-                        row(inputs_left, Spacer(width=10), inputs_right))
-        layout = column(row(inputs, Spacer(width=10), self.primary.plot),
-                        self.secondary.plot,)
-
-        # Load and display the data
-        self.primary.callback(None, None, None)
+        layout = column(self.primary.layout(), self.secondary.layout())
 
         # Go!
         curdoc().add_root(layout)
