+++ delicatessen/modules/__init__.py
@@ -1,0 +1,2 @@
+from .base import BasePlot
+from . import base, TESSGaiaMini
