+++ delicatessen/modules/base.py
@@ -1,0 +1,15 @@
+# Third-party
+import numpy as np
+from bokeh.models import ColumnDataSource, Div
+from bokeh.plotting import figure
+
+
+class BasePlot:
+    def __init__(self, primary_plot):
+        self.primary_plot = primary_plot
+
+    def callback(self, attr, old, new):
+        pass
+
+    def layout(self):
+        return Div()
